changed test data to fit new model, restructured code
