--- src/main/java/com/lecture/coordinator/model/Semester.java
@@ -1,5 +1,0 @@
-package com.lecture.coordinator.model;
-
-public enum Semester {
-    SS, WS
-}
