+++ src/main/java/com/lecture/coordinator/model/enums/CourseType.java
@@ -1,4 +1,4 @@
-package com.lecture.coordinator.model;
+package com.lecture.coordinator.model.enums;
 
 public enum CourseType {
     VO,
