+++ src/main/java/com/lecture/coordinator/model/enums/Semester.java
@@ -1,0 +1,5 @@
+package com.lecture.coordinator.model.enums;
+
+public enum Semester {
+    SS, WS
+}
