+++ src/main/java/com/lecture/coordinator/repositories/TimingRepository.java
@@ -2,12 +2,13 @@ package com.lecture.coordinator.repositories;
 
 import com.lecture.coordinator.model.Course;
 import com.lecture.coordinator.model.Room;
+import com.lecture.coordinator.model.RoomTable;
 import com.lecture.coordinator.model.Timing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface TimingRepository extends JpaRepository<Timing, Long> {
-    public List<Timing> findAllByRoom(Room room);
+    public List<Timing> findAllByRoomTable(RoomTable roomTable);
     public List<Timing> findAllByCourse(Course course);
 }
