+++ src/main/java/com/lecture/coordinator/services/TimeTableService.java
@@ -1,6 +1,7 @@
 package com.lecture.coordinator.services;
 
 import com.lecture.coordinator.model.*;
+import com.lecture.coordinator.model.enums.Semester;
 import com.lecture.coordinator.repositories.TimeTableRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
